Add a command for fixing public menu URI’s
Sometimes, menu items are created with a public URI, e.g. `/melp`. This causes
problems when checking for the current item, which uses the internal URI. Stuff like breadcrumbs
will fail to render correctly because this.
The added command fixes this: it checks if there are any menu items with a public URI and replaces
that with the corresponding internal URI.